% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlapi_S3.R
\name{fit_transform}
\alias{fit_transform}
\alias{fit_transform.Matrix}
\alias{fit_transform.matrix}
\title{Fit model to the data, then transforms data}
\usage{
fit_transform(x, model, y = NULL, ...)

\method{fit_transform}{Matrix}(x, model, y = NULL, ...)

\method{fit_transform}{matrix}(x, model, y = NULL, ...)
}
\arguments{
\item{x}{A matrix like object, should \bold{inherit from \code{Matrix} or \code{matrix}}}

\item{model}{instance of class \code{estimator} which should implement method
with signature \code{$fit(x, ...)}}

\item{y}{\code{NULL} by default. Optional response variable for supervised models.
Should \bold{inherit from \code{vector} \code{Matrix} or \code{matrix}}.
See documentation for corresponding models.}

\item{...}{additional data/model dependent arguments to downstream functions.}
}
\value{
Transformed version of the \code{x}
}
\description{
Generic function to fit transformers (inherits from \link{mlapiTransformation})
}
