% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\docType{data}
\name{mlapiDecomposition}
\alias{mlapiDecomposition}
\alias{mlapiDecomposition}
\alias{mlapiDecompositionOnline}
\title{Base abstract class for all decompositions}
\format{\code{R6Class} object.}
\usage{
mlapiDecomposition

mlapiDecompositionOnline
}
\description{
Base class for all \bold{decompositions} which are methods which can decompose matrix into
2 low-dimensional matrices \code{x = f(A, B)}.
(Think of this Latent Dirichlet Allocation, Non-negative Matrix Factorization, etc).
It iherits from \link{mlapiTransformation} and additionally requires to implement \code{components} member.

Base class for all \bold{decompositions} which are methods which can decompose matrix into
2 low-dimensional matrices \code{x = f(A, B)} \bold{incrementally}.
It iherits from \link{mlapiDecomposition} and additionally requires
to implement \code{partial_fit} method which can learn \code{components} incrementally.
}
\section{Fields}{

\describe{
\item{\code{components}}{features embeddings. So if matrix is decomposed in a form \code{x = f(A, B)} where
X = n\*m, A = n\*k, B = k\*m them \code{B = components}}

\item{\code{components}}{features embeddings. So if matrix is decomposed in a form \code{x = f(A, B)} where
X = n\*m, A = n\*k, B = k\*m them \code{B = components}}
}}

\section{Methods}{

\describe{
  \item{\code{$fit_transform(x, y = NULL, ...)}}{}
  \item{\code{$transform(x, ...)}}{Performs transformation of the new data (after model was trained)}
}


\describe{
  \item{\code{$fit_transform(x, y = NULL, ...)}}{}
  \item{\code{$partial_fit(x, y = NULL, ...)}}{}
  \item{\code{$transform(x, ...)}}{Performs transformation of the new data (after model was trained)}
}
}

\section{Arguments}{

\describe{
 \item{x}{A matrix like object, should \bold{inherit from \code{Matrix} or \code{matrix}}.
 Allowed classes should be defined in child classes.}
 \item{y}{\code{NULL}. Optional taget variable. Usually this should be \code{NULL}.
 There few cases when it could be used.}
 \item{...}{additional parameters \bold{with default values}}
}


\describe{
 \item{x}{A matrix like object, should \bold{inherit from \code{Matrix} or \code{matrix}}.
 Allowed classes should be defined in child classes.}
 \item{y}{\code{NULL}. Optional taget variable. Usually this should be \code{NULL}.
 There few cases when it could be used.}
 \item{...}{additional parameters \bold{with default values}}
}
}

\keyword{datasets}
