% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\docType{data}
\name{mlapiEstimator}
\alias{mlapiEstimator}
\title{Base abstract class for all classification/regression models}
\format{\code{R6Class} object.}
\usage{
mlapiEstimator
}
\description{
Base class for all estimators.
Defines minimal set of members and methods(with signatires) which have to be implemented in child classes.
}
\section{Methods}{

\describe{
  \item{\code{$fit(x, y, ...)}}{}
  \item{\code{$predict(x, ...)}}{Makes predictions on new data (after model was trained)}
}
}

\section{Arguments}{

\describe{
 \item{x}{A matrix like object, should \bold{inherit from \code{Matrix} or \code{matrix}}.
 Allowed classes should be defined in child classes.}
 \item{y}{target - usually \code{vector}, but also can be a matrix like object.
 Allowed classes should be defined in child classes.}
 \item{...}{additional parameters \bold{with default values}}
}
}

\keyword{datasets}
