% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\docType{data}
\name{mlapiTransformation}
\alias{mlapiTransformation}
\title{Base abstract class for all transformations}
\format{\code{R6Class} object.}
\usage{
mlapiTransformation
}
\description{
Base class for all online transformations.
}
\section{Methods}{

\describe{
  \item{\code{$fit_transform(x, y = NULL, ...)}}{}
  \item{\code{$transform(x, ...)}}{Performs transformation of the new data (after model was trained)}
}
}

\section{Arguments}{

\describe{
 \item{x}{A matrix like object, should \bold{inherit from \code{Matrix} or \code{matrix}}.
 Allowed classes should be defined in child classes.}
 \item{y}{\code{NULL}. Optional taget variable. Usually this should be \code{NULL}.
 There few cases when it could be used.}
 \item{...}{additional parameters \bold{with default values}}
}
}

\keyword{datasets}
