% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformer.R
\docType{data}
\name{mlapiTransformationOnline}
\alias{mlapiTransformationOnline}
\title{Base abstract class for all transformations
which can be \bold{trained incremendally} (online)}
\format{\code{R6Class} object.}
\usage{
mlapiTransformationOnline
}
\description{
Base class for all online transformations. This class inherits from \link{mlapiTransformation} and
additionally requires to implement \code{$partial_fit(x, y, ...)} method. Idea is that user can pass
\code{x, y} in chunks and model will be updated/refined incrementally.
}
\section{Methods}{

\describe{
  \item{\code{$fit_transform(x, y = NULL, ...)}}{}
  \item{\code{$transform(x, ...)}}{Performs transformation of the new data (after model was trained)}
}
}

\section{Arguments}{

\describe{
 \item{x}{A matrix like object, should \bold{inherit from \code{Matrix} or \code{matrix}}.
 Allowed classes should be defined in child classes.}
 \item{y}{\code{NULL}. Optional taget variable. Usually this should be \code{NULL}.
 There few cases when it could be used.}
 \item{...}{additional parameters \bold{with default values}}
}
}

\keyword{datasets}
