% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlapi_S3.R
\name{predict}
\alias{predict}
\alias{predict.matrix}
\alias{predict.Matrix}
\title{Makes predictions on new data using pre-trained model}
\usage{
\method{predict}{matrix}(object, model, ...)

\method{predict}{Matrix}(object, model, ...)
}
\arguments{
\item{object}{\bold{ = x} in other methods.
A matrix like object, should \bold{inherit from \code{Matrix} or \code{matrix}}}

\item{model}{object which \bold{inherits} class \link{mlapiEstimation} which
implements method \code{model$predict(x, ...)}}

\item{...}{additional data/model dependent arguments to downstream functions}
}
\description{
Makes predictions on new data using pre-trained \code{model} (inherits from \link{mlapiEstimation})
}
